requirements.txt
# Core ML/Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Network Traffic Analysis
scapy>=2.5.0
pyshark>=0.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
click>=8.1.0
rich>=13.0.0

# Jupyter (for local notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
